requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
motor==3.1.2
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
